requirements.txt
asgiref~=3.8.1
Django~=5.0.8
django-browser-reload~=1.13.0
python-decouple~=3.8
sqlparse~=0.5.1
logging
pyjwt
requests